# backend/openapi.yaml
openapi: 3.1.0
info:
  title: Trivela API
  version: 0.1.0
  description: |
    REST API for the Trivela campaign and rewards platform. Handles campaign metadata,
    health checks, and Soroban RPC configuration for the frontend.

    See the [API Migration Guide](../docs/API_MIGRATION.md) for v0 → v1 breaking changes,
    endpoint changelog, deprecation policy, and migration examples with curl snippets.
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: http://localhost:3001
    description: Local development server
  - url: https://api.trivela.example.com
    description: Production server
...
security:
  - ApiKeyAuth: []
  - BearerAuth: []
paths:
...
  /api/v1/sponsored-accounts:
    post:
      summary: Sponsor a new Stellar account (CAP-33)
      tags: [SponsoredAccounts]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [address]
              properties:
                address:
                  type: string
                  description: Stellar G-address to sponsor
                accountType:
                  type: string
                  enum: [stellar, smart_wallet]
                  default: stellar
                trustlineAsset:
                  type: string
                  description: '"CODE:ISSUER" trustline to add within the sponsorship envelope'
      responses:
        '201':
          description: Sponsorship created; returns partially-signed transaction XDR
        '202':
          description: Sponsorship tracked (SPONSOR_SECRET_KEY not configured)
        '400':
          description: Invalid address or asset format
        '409':
          description: Address already has an active sponsorship
  /api/v1/sponsored-accounts/{address}:
    get:
      summary: Get sponsorship status for an address
      tags: [SponsoredAccounts]
      parameters:
        - in: path
          name: address
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Sponsorship record
        '404':
          description: Not found
  /api/v1/sponsored-accounts/{address}/revoke:
    post:
      summary: Revoke reserve sponsorship (transfer back to account)
      tags: [SponsoredAccounts]
      parameters:
        - in: path
          name: address
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Revocation XDR built; or tracked-only if key not configured
        '404':
          description: No active sponsorship for this address

  # ── #548: Claimable balances ───────────────────────────────────────────────
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for write operations
    BearerAuth:
      type: http
      scheme: bearer
      description: Bearer token authentication
...